requirements.txt
kubernetes>=28.1.0
PyYAML>=6.0
rich>=13.0.0